+++ lib/src/clean_structure_creators/clean_i_creators.dart
@@ -1,0 +1,16 @@
+import 'dart:io';
+
+abstract class CleanISSLCreator {
+  Future<void> create();
+}
+
+abstract class IDirectoryCreator {
+  Future<bool> createDirectories();
+  Directory get coreDir;
+  Directory get featuresDir;
+  Directory get assetsDir;
+}
+
+abstract class IFileCreator {
+  Future<void> createNecessaryFiles();
+}